240716 dsafj112

+++ src/main/java/com/boot/dto/CorpInfoDTO.java
@@ -1,5 +1,6 @@
 package com.boot.dto;
 
+import java.math.BigInteger;
 import java.sql.Timestamp;
 import java.util.List;
 
@@ -17,7 +18,7 @@ public class CorpInfoDTO {
 	private String corp_address;
 	private int corpInfo_Sal;
 	private int corpInfo_empNo;
-	private int corpInfo_money;
+	private BigInteger corpInfo_money;
 	private String corpInfo_intro;
 	private Timestamp corpInfo_birth;
 	private List<String> corp_keyword; // 키워드 리스트 추가
